Refactor project-geometries to accept layerGroups as arg
