+++ app/components/project-geometries.js
@@ -8,127 +8,13 @@ import { developmentSiteLayer } from './project-geometries/types/development-sit
 import { projectAreaLayer } from './project-geometries/types/project-area';
 import projectGeometryIcons from '../utils/project-geom-icons';
 
-const mapEditingLayerGroups = {
-  'layer-groups': [
-    {
-      id: 'tax-lots',
-      visible: true,
-      layers: [
-        {
-          highlightable: true,
-          clickable: false,
-          tooltipable: true,
-          tooltipTemplate: '{{address}} (BBL: {{bbl}})',
-        },
-        {},
-        { style: { layout: { 'text-field': '{lot}' } } },
-        {
-          style: {
-            id: 'block-labels',
-            type: 'symbol',
-            source: 'pluto',
-            'source-layer': 'block-centroids',
-            minzoom: 14,
-            maxzoom: 24,
-            layout: {
-              'text-field': '{block}',
-              'text-font': [
-                'Open Sans Bold',
-                'Arial Unicode MS Regular',
-              ],
-              'text-size': 22,
-            },
-            paint: {
-              'text-halo-color': 'rgba(255, 255, 255, 0.5)',
-              'text-halo-width': 1,
-              'text-color': 'rgba(121, 121, 121, 1)',
-              'text-halo-blur': 0,
-              'text-opacity': {
-                stops: [
-                  [
-                    14,
-                    0,
-                  ],
-                  [
-                    15,
-                    1,
-                  ],
-                ],
-              },
-            },
-          },
-        },
-      ],
-    },
-    {
-      id: 'zoning-districts',
-      visible: true,
-      layers: [
-        {
-          highlightable: false,
-          tooltipable: false,
-          style: {
-            paint: {
-              'fill-opacity': 0,
-            },
-          },
-        },
-        {
-          tooltipable: false,
-          style: {
-            paint: {
-              'line-opacity': 0.05,
-            },
-          },
-        },
-        {
-          style: {
-            paint: {
-              'text-opacity': 0,
-            },
-          },
-        },
-      ],
-    },
-    {
-      id: 'street-centerlines',
-      visible: true,
-      layers: [
-        {},
-        {
-          before: 'place_country_major',
-          style: {
-            id: 'citymap-street-centerlines-line',
-            type: 'line',
-            source: 'digital-citymap',
-            'source-layer': 'street-centerlines',
-            metadata: {
-              'nycplanninglabs:layergroupid': 'street-centerlines',
-            },
-            minzoom: 13,
-            paint: {
-              'line-dasharray': [
-                5,
-                3,
-              ],
-              'line-color': 'rgba(193, 193, 193, 1)',
-              'line-width': 0.5,
-            },
-          },
-        },
-      ],
-    },
-  ],
-};
-
+// it:
+// renders the map
+// sets up a drawable map
+// invokes the types-/mode-renderer with drawable map
+// and conditionally loads the map legend
 @tagName('')
 export default class ProjectGeometryEditComponent extends Component {
-  constructor(...args) {
-    super(...args);
-
-    this.loadLayerGroups();
-  }
-
   // required
   @argument
   model;
@@ -140,6 +26,10 @@ export default class ProjectGeometryEditComponent extends Component {
   // required
   @argument
   mode;
+
+  // required
+  @argument
+  layerGroups;
 
   @service
   notificationMessages;
@@ -156,16 +46,9 @@ export default class ProjectGeometryEditComponent extends Component {
   projectAreaLayer = projectAreaLayer;
 
   /* ----------  General Map  ---------- */
-  layerGroups = null;
-
   @computed('lat', 'lng')
   get center() {
     return [this.get('lat'), this.get('lng')];
-  }
-
-  @action
-  rollbackChanges() {
-    this.get('model').rollbackAttributes();
   }
 
   @action
@@ -196,18 +79,5 @@ export default class ProjectGeometryEditComponent extends Component {
     ];
 
     basemapLayersToHide.forEach(layer => map.removeLayer(layer));
-  }
-
-  // TODO: This function seems awkward. layerGroups.layerGroups?
-  loadLayerGroups() {
-    const store = this.get('store');
-    store.query('layer-group', mapEditingLayerGroups).then((layerGroups) => {
-      const { meta } = layerGroups;
-
-      this.set('layerGroups', {
-        layerGroups,
-        meta,
-      });
-    });
   }
 }
