+++ app/templates/components/project-geometries.hbs
@@ -1,3 +1,5 @@
+{{log layerGroups}}
+
 <div class="cell sidebar map-sidebar">
   {{!-- EMBER WORMHOLE CONTAINER --}}
   {{!-- The geometry type component will fill this div with markup once invoked --}}
