+++ app/templates/components/project-geometries/map-legend.hbs
@@ -1,0 +1,1 @@
+{{yield}}
