+++ app/templates/projects/edit/geometry-edit.hbs
@@ -1,5 +1,8 @@
+{{!-- {{log model.layerGroups}} --}}
+
 {{project-geometries
-  model=model
+  model=model.project
+  layerGroups=model.layerGroups
   type=type
   mode=mode}}
