+++ tests/integration/components/project-geometries/map-legend-test.js
@@ -1,0 +1,26 @@
+import { module, test } from 'qunit';
+import { setupRenderingTest } from 'ember-qunit';
+import { render } from '@ember/test-helpers';
+import hbs from 'htmlbars-inline-precompile';
+
+module('Integration | Component | project-geometries/map-legend', function(hooks) {
+  setupRenderingTest(hooks);
+
+  test('it renders', async function(assert) {
+    // Set any properties with this.set('myProperty', 'value');
+    // Handle any actions with this.set('myAction', function(val) { ... });
+
+    await render(hbs`{{project-geometries/map-legend}}`);
+
+    assert.equal(this.element.textContent.trim(), '');
+
+    // Template block usage:
+    await render(hbs`
+      {{#project-geometries/map-legend}}
+        template block text
+      {{/project-geometries/map-legend}}
+    `);
+
+    assert.equal(this.element.textContent.trim(), 'template block text');
+  });
+});
